requirements.txt
uvicorn==0.22.0
fastapi==0.96.0
fastapi-offline==1.5.3
pathlib==1.0.1
pyannote.audio==2.1.1
pydub==0.25.1
python-multipart==0.0.6
scikit-learn==1.2.2
numpy==1.23.5
tensorflow==2.12.0
torch==1.13.1
torchaudio==0.13.1